Add lock to prevent updateGui() from running while pages are changing
There was previously a race condition where if a user pressed a button
that eventually calls updateGui() after the active phase index was
changed but before the view pointer is updated, updateGui() would update
the old view instead of the new one.  To fix this, a variable is set to
true during this period and updateGui() does not run while the variable
is true.