Constrain POI button bottom to information line top
It was previously constrained to the bottom of the view, assuming that the
margin would be enough for it to avoid the information line text. That's not the
case when the information line wraps to two lines, or if the user is using a
large font (since the text size is specified in sp, but the margin is specified
in dp). This ties the POI button to the information line TextView. The price is
that the POI button on a two-info-line screen (e.g., Canadian Parliamentary,
Leader of the Opposition) is now smaller than on a one-info-line screen (most
others).